+++ src/game/clock.cpp
@@ -18,7 +18,7 @@ regulate(Clock& clock)
     u32 time_needed_us = (clock.n_instructions * CLOCK_RATE_NS) / 1000u;
     u64 now_us = Time::get_us();
 
-    u32 time_taken_us = now_us - clock.last_reset_us;
+    u32 time_taken_us = (u32)(now_us - clock.last_reset_us);
 
     clock.last_reset_us = now_us;
     clock.n_instructions = 0;
